requirements.txt
Flask==3.0.3
tensorflow==2.17.0
torch==2.0.1
torchvision==0.15.2
numpy==1.26.4
Pillow==10.4.0
